pyproject.toml
[project]
name = "medium-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "celery>=5.5.3",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "jose>=1.0.0",
    "passlib[context]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "starlette-admin>=0.15.1",
]
